Add type_ module
